+++ src/main/java/com/uber/jenkins/phabricator/coverage/JacocoCoverageProvider.java
@@ -83,8 +83,8 @@ public class JacocoCoverageProvider extends CoverageProvider {
         float classCoverage = coverageResult.getClassCoverage().getPercentageFloat();
         float lineCoverage = coverageResult.getLineCoverage().getPercentageFloat();
         float branchCoverage = coverageResult.getBranchCoverage().getPercentageFloat();
-        float linesCovered = coverageResult.getLineCoverage().getCovered();
-        float linesTested = coverageResult.getLineCoverage().getTotal();
+        long linesCovered = coverageResult.getLineCoverage().getCovered();
+        long linesTested = coverageResult.getLineCoverage().getTotal();
 
         return new CodeCoverageMetrics(
                 PERCENTAGE_UNAVAILABLE,
